Add exec. module for annotating ref. samples CSV
